[DataGrid] Merge useGridVirtualColumns into useGridVirtualRows
