+++ packages/grid/_modules_/grid/hooks/features/virtualization/index.ts
@@ -1,3 +1,2 @@
-export * from './useGridVirtualColumns';
 export * from './useGridVirtualRows';
 export * from './renderingState';
