+++ packages/grid/_modules_/grid/hooks/features/virtualization/useGridVirtualRows.ts
@@ -4,7 +4,12 @@ import { GridApiRef } from '../../../models/api/gridApiRef';
 import { GridVirtualizationApi } from '../../../models/api/gridVirtualizationApi';
 import { GridCellIndexCoordinates } from '../../../models/gridCell';
 import { GridScrollParams } from '../../../models/params/gridScrollParams';
-import { GridRenderContextProps, GridRenderRowProps } from '../../../models/gridRenderContextProps';
+import {
+  GridRenderContextProps,
+  GridRenderRowProps,
+  GridRenderColumnsProps,
+} from '../../../models/gridRenderContextProps';
+import { GridContainerProps } from '../../../models/gridContainerProps';
 import { isDeepEqual } from '../../../utils/utils';
 import { useEnhancedEffect } from '../../../utils/material-ui-utils';
 import { optionsSelector } from '../../utils/optionsSelector';
@@ -21,9 +26,9 @@ import { useNativeEventListener } from '../../root/useNativeEventListener';
 import { useLogger } from '../../utils/useLogger';
 import { useGridScrollFn } from '../../utils/useGridScrollFn';
 import { InternalRenderingState } from './renderingState';
-import { useGridVirtualColumns } from './useGridVirtualColumns';
 import { gridDensityRowHeightSelector } from '../density/densitySelector';
 import { scrollStateSelector } from './renderingStateSelector';
+import { useGridApiEventHandler } from '../../root/useGridApiEventHandler';
 
 // Logic copied from https://www.w3.org/TR/wai-aria-practices/examples/listbox/js/listbox.js
 // Similar to https://developer.mozilla.org/en-US/docs/Web/API/Element/scrollIntoView
@@ -40,6 +45,28 @@ function scrollIntoView(dimensions) {
   return undefined;
 }
 
+// Uses binary search to avoid looping through all possible positions
+function getIdxFromScroll(
+  offset: number,
+  positions: number[],
+  sliceStart = 0,
+  sliceEnd = positions.length,
+): number {
+  if (positions.length <= 0) {
+    return -1;
+  }
+
+  if (sliceStart >= sliceEnd) {
+    return sliceStart;
+  }
+
+  const pivot = sliceStart + Math.floor((sliceEnd - sliceStart) / 2);
+  const itemOffset = positions[pivot];
+  return offset <= itemOffset
+    ? getIdxFromScroll(offset, positions, sliceStart, pivot)
+    : getIdxFromScroll(offset, positions, pivot + 1, sliceEnd);
+}
+
 export const useGridVirtualRows = (apiRef: GridApiRef): void => {
   const logger = useLogger('useGridVirtualRows');
   const colRef = apiRef.current.columnHeadersElementRef!;
@@ -53,9 +80,11 @@ export const useGridVirtualRows = (apiRef: GridApiRef): void => {
   const totalRowCount = useGridSelector(apiRef, gridRowCountSelector);
   const visibleColumns = useGridSelector(apiRef, visibleGridColumnsSelector);
   const columnsMeta = useGridSelector(apiRef, gridColumnsMetaSelector);
+  const renderedColRef = React.useRef<GridRenderColumnsProps | null>(null);
+  const containerPropsRef = React.useRef<GridContainerProps | null>(null);
+  const lastScrollLeftRef = React.useRef<number>(0);
 
   const [scrollTo] = useGridScrollFn(renderingZoneRef, colRef);
-  const [renderedColRef, updateRenderedCols] = useGridVirtualColumns(options, apiRef);
 
   const setRenderingState = React.useCallback(
     (newState: Partial<InternalRenderingState>) => {
@@ -122,15 +151,101 @@ export const useGridVirtualRows = (apiRef: GridApiRef): void => {
 
   const reRender = React.useCallback(() => {
     const renderingState = getRenderingState();
-    const hasChanged = setRenderingState({
-      renderContext: renderingState,
-      renderedSizes: apiRef.current.state.containerSizes,
-    });
+    const hasChanged = setRenderingState({ renderContext: renderingState });
     if (hasChanged) {
       logger.debug('reRender: trigger rendering');
       forceUpdate();
     }
-  }, [apiRef, getRenderingState, logger, forceUpdate, setRenderingState]);
+  }, [getRenderingState, logger, forceUpdate, setRenderingState]);
+
+  const getColumnIdxFromScroll = React.useCallback(
+    (left: number) => getIdxFromScroll(left, columnsMeta.positions),
+    [columnsMeta.positions],
+  );
+
+  const getColumnFromScroll = React.useCallback(
+    (left: number) => {
+      if (!visibleColumns.length) {
+        return null;
+      }
+      return visibleColumns[getColumnIdxFromScroll(left)];
+    },
+    [getColumnIdxFromScroll, visibleColumns],
+  );
+
+  const updateRenderedCols = React.useCallback(
+    (containerProps: GridContainerProps | null, scrollLeft: number) => {
+      if (!containerProps) {
+        return false;
+      }
+
+      containerPropsRef.current = containerProps;
+      const windowWidth = containerProps.windowSizes.width;
+
+      lastScrollLeftRef.current = scrollLeft;
+      logger.debug(
+        `GridColumns from ${getColumnFromScroll(scrollLeft)?.field} to ${
+          getColumnFromScroll(scrollLeft + windowWidth)?.field
+        }`,
+      );
+      const firstDisplayedIdx = getColumnIdxFromScroll(scrollLeft);
+      const lastDisplayedIdx = getColumnIdxFromScroll(scrollLeft + windowWidth);
+      const prevFirstColIdx = renderedColRef?.current?.firstColIdx || 0;
+      const prevLastColIdx = renderedColRef?.current?.lastColIdx || 0;
+      const columnBuffer = options.columnBuffer;
+      const tolerance = columnBuffer > 1 ? columnBuffer - 1 : columnBuffer; // Math.floor(columnBuffer / 2);
+      const diffFirst = Math.abs(firstDisplayedIdx - tolerance - prevFirstColIdx);
+      const diffLast = Math.abs(lastDisplayedIdx + tolerance - prevLastColIdx);
+      logger.debug(`Column buffer: ${columnBuffer}, tolerance: ${tolerance}`);
+      logger.debug(`Previous values  => first: ${prevFirstColIdx}, last: ${prevLastColIdx}`);
+      logger.debug(
+        `Current displayed values  => first: ${firstDisplayedIdx}, last: ${lastDisplayedIdx}`,
+      );
+      logger.debug(`Difference with first: ${diffFirst} and last: ${diffLast} `);
+
+      const lastVisibleColIdx = visibleColumns.length > 0 ? visibleColumns.length - 1 : 0;
+      const firstColIdx =
+        firstDisplayedIdx - columnBuffer >= 0 ? firstDisplayedIdx - columnBuffer : 0;
+      const newRenderedColState = {
+        leftEmptyWidth: columnsMeta.positions[firstColIdx],
+        rightEmptyWidth: 0,
+        firstColIdx,
+        lastColIdx:
+          lastDisplayedIdx + columnBuffer >= lastVisibleColIdx
+            ? lastVisibleColIdx
+            : lastDisplayedIdx + columnBuffer,
+      };
+
+      if (apiRef.current.state.scrollBar.hasScrollX) {
+        newRenderedColState.rightEmptyWidth =
+          columnsMeta.totalWidth -
+          columnsMeta.positions[newRenderedColState.lastColIdx] -
+          visibleColumns[newRenderedColState.lastColIdx].computedWidth;
+      } else if (!options.disableExtendRowFullWidth) {
+        newRenderedColState.rightEmptyWidth =
+          apiRef.current.state.viewportSizes.width - columnsMeta.totalWidth;
+      }
+
+      if (!isDeepEqual(newRenderedColState, renderedColRef.current)) {
+        renderedColRef.current = newRenderedColState;
+        logger.debug('New columns state to render', newRenderedColState);
+        return true;
+      }
+      logger.debug(`No rendering needed on columns`);
+      return false;
+    },
+    [
+      apiRef,
+      columnsMeta.positions,
+      columnsMeta.totalWidth,
+      getColumnFromScroll,
+      getColumnIdxFromScroll,
+      logger,
+      options.columnBuffer,
+      options.disableExtendRowFullWidth,
+      visibleColumns,
+    ],
+  );
 
   const updateViewport = React.useCallback(
     (forceReRender = false) => {
@@ -198,6 +313,7 @@ export const useGridVirtualRows = (apiRef: GridApiRef): void => {
     ],
   );
 
+  // TODO move to useGridScroll, it's the same regardless of whether virtualization is on or not
   const scrollToIndexes = React.useCallback(
     (params: Partial<GridCellIndexCoordinates>) => {
       if (totalRowCount === 0 || visibleColumns.length === 0) {
@@ -285,6 +401,7 @@ export const useGridVirtualRows = (apiRef: GridApiRef): void => {
     }
   }, [windowRef, apiRef, setGridState, forceUpdate]);
 
+  // TODO move to useGridScroll, it's the same regardless of whether virtualization is on or not
   const scroll = React.useCallback(
     (params: Partial<GridScrollParams>) => {
       if (windowRef.current && params.left != null && colRef.current) {
@@ -350,20 +467,14 @@ export const useGridVirtualRows = (apiRef: GridApiRef): void => {
   ]);
 
   React.useEffect(() => {
-    if (
-      gridState.containerSizes !== gridState.rendering.renderedSizes &&
-      apiRef.current.updateViewport
-    ) {
-      logger.debug(`gridState.containerSizes updated, updating viewport. `);
-      apiRef.current.updateViewport(true);
-    }
-  }, [apiRef, gridState.containerSizes, gridState.rendering.renderedSizes, logger]);
-
-  React.useEffect(() => {
     if (apiRef.current.updateViewport) {
       logger.debug(`totalRowCount has changed to ${totalRowCount}, updating viewport.`);
       apiRef.current.updateViewport(true);
     }
+
+    return () => {
+      clearTimeout(scrollingTimeout.current);
+    };
   }, [
     logger,
     totalRowCount,
@@ -372,12 +483,6 @@ export const useGridVirtualRows = (apiRef: GridApiRef): void => {
     gridState.containerSizes,
     apiRef,
   ]);
-
-  React.useEffect(() => {
-    return () => {
-      clearTimeout(scrollingTimeout.current);
-    };
-  }, []);
 
   const preventScroll = React.useCallback((event: any) => {
     event.target.scrollLeft = 0;
@@ -397,4 +502,12 @@ export const useGridVirtualRows = (apiRef: GridApiRef): void => {
     'scroll',
     preventScroll,
   );
+
+  const resetRenderedColState = React.useCallback(() => {
+    logger.debug('Clearing previous renderedColRef');
+    renderedColRef.current = null;
+  }, [logger, renderedColRef]);
+
+  useGridApiEventHandler(apiRef, GridEvents.columnsChange, resetRenderedColState);
+  useGridApiEventHandler(apiRef, GridEvents.debouncedResize, resetRenderedColState);
 };
